move np generate to v1beta1
