+++ pkg/apis/softwarecomposition/v1beta1/networkpolicy/networkpolicy_test.go
@@ -3,7 +3,7 @@ package networkpolicy
 import (
 	"testing"
 
-	"github.com/kubescape/storage/pkg/apis/softwarecomposition"
+	softwarecomposition "github.com/kubescape/storage/pkg/apis/softwarecomposition/v1beta1"
 	"github.com/stretchr/testify/assert"
 	corev1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
