+++ pkg/registry/file/generatednetworkpolicy.go
@@ -7,8 +7,8 @@ import (
 
 	"github.com/kubescape/go-logger"
 	"github.com/kubescape/go-logger/helpers"
-	"github.com/kubescape/storage/pkg/apis/softwarecomposition"
-	"github.com/kubescape/storage/pkg/apis/softwarecomposition/networkpolicy"
+	softwarecomposition "github.com/kubescape/storage/pkg/apis/softwarecomposition/v1beta1"
+	networkpolicy "github.com/kubescape/storage/pkg/apis/softwarecomposition/v1beta1/networkpolicy"
 	"go.opentelemetry.io/otel"
 	"go.opentelemetry.io/otel/attribute"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
